requirements.txt
gradio>=3.50.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
imageio>=2.15.0
deap>=1.3.0
matplotlib>=3.5.0
plotly>=5.10.0
pandas>=1.5.0
psutil>=5.9.0
Pillow>=9.0.0
opencv-python>=4.5.0
PyYAML>=6.0